+++ kernel_tuner/interface.py
@@ -352,7 +352,7 @@ def tune_kernel(kernel_name, kernel_source, problem_size, arguments, tune_params
     objective, objective_higher_is_better = get_objective_defaults(objective, objective_higher_is_better)
 
     # check for forbidden names in tune parameters
-    util.check_tune_params_list(tune_params)
+    util.check_tune_params_list(tune_params, observers)
 
     # check whether block_size_names are used as expected
     util.check_block_size_params_names_list(block_size_names, tune_params)
@@ -415,7 +415,8 @@ def tune_kernel(kernel_name, kernel_source, problem_size, arguments, tune_params
         strategy = brute_force
 
     # select the runner for this job based on input
-    selected_runner = SimulationRunner if simulation_mode is True else SequentialRunner
+    selected_runner = SimulationRunner if simulation_mode else SequentialRunner
+    tuning_options.simulated_time = 0
     runner = selected_runner(kernelsource, kernel_options, device_options, iterations, observers)
 
     # the user-specified function may or may not have an optional atol argument;